star

+++ db/migrate/20190821044553_create_reviews.rb
@@ -1,7 +1,7 @@
 class CreateReviews < ActiveRecord::Migration[5.0]
   def change
     create_table :reviews do |t|
-      t.integer :rate
+      t.float :rate
       t.text :review
       t.integer :shop_id,    foreign_key: true
       t.integer :user_id,    foreign_key: true
